Created rendering utilities as separate entity
