+++ graphics/CMakeLists.txt
@@ -14,7 +14,9 @@ add_library(graphics
 ## For GLEW
 # GLEW is a required package, so report config failure if not found
 find_package(GLEW REQUIRED)
-target_link_libraries(graphics PUBLIC GLEW::GLEW)
+find_package(CGAL REQUIRED)
+include_directories( $ENV{EIGEN3_INCLUDE_DIR} )
+target_link_libraries(graphics geom_utils GLEW::GLEW ${CGAL_LIBRARIES})
 
 message(${GLEW_INCLUDE_DIRS})
 message(${GLEW_LIBRARIES})
